#38 example sequences and docs
