+++ control/test/config/livex.cfg
@@ -45,7 +45,7 @@ ki_default = 0.02
 kd_default = 0
 
 # Include ALL information sent via TCP in saved data, which includes additional PID information
-pid_debug = 1
+pid_debug = 0
 # Use of a mocked modbus client, exclusively for testing without a real furnace
 # Mock client has all registers, but only heater A will do anything
 use_mock_client = 0
